# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen task/todo list",
        "title": "Task Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/todos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all todos for the authenticated user with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Get all todos",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "personal",
                            "work",
                            "shopping",
                            "health",
                            "other"
                        ],
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todos retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new todo item for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "parameters": [
                    {
                        "description": "Create Todo Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTodoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    }
                }
            }
        },
        "/todos/public/{id}": {
            "get": {
                "description": "Get detail of a todo (and subtasks) using its public ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Get todo by public ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Public ID of the Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific todo by ID for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Get a todo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid todo ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a specific todo by ID for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Update a todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Todo Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateTodoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or todo ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific todo by ID for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or todo ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.TodoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreateTodoRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "category": {
                    "enum": [
                        "personal",
                        "work",
                        "shopping",
                        "health",
                        "other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ],
                    "example": "personal"
                },
                "description": {
                    "type": "string",
                    "example": "Need to buy milk, eggs, and bread"
                },
                "due_date": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "priority": {
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Priority"
                        }
                    ],
                    "example": "medium"
                },
                "title": {
                    "type": "string",
                    "example": "Buy groceries"
                }
            }
        },
...
        "handlers.TodoResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "handlers.UpdateTodoRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "enum": [
                        "personal",
                        "work",
                        "shopping",
                        "health",
                        "other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ],
                    "example": "work"
                },
                "description": {
                    "type": "string",
                    "example": "Need to buy milk, eggs, and bread"
                },
                "due_date": {
                    "type": "string",
                    "example": "2024-12-31T23:59:59Z"
                },
                "priority": {
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Priority"
                        }
                    ],
                    "example": "high"
                },
                "status": {
                    "enum": [
                        "pending",
                        "in_progress",
                        "completed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Status"
                        }
                    ],
                    "example": "completed"
                },
                "title": {
                    "type": "string",
                    "example": "Buy groceries"
                }
            }
        },
        "models.Category": {
            "type": "string",
            "enum": [
                "work",
                "personal",
                "shopping",
                "health",
                "other"
            ],
            "x-enum-varnames": [
                "CategoryWork",
                "CategoryPersonal",
                "CategoryShopping",
                "CategoryHealth",
                "CategoryOther"
            ]
        },
        "models.Priority": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "PriorityLow",
                "PriorityMedium",
                "PriorityHigh"
            ]
        },
        "models.Status": {
            "type": "string",
            "enum": [
                "todo",
                "inprogress",
                "done"
            ],
            "x-enum-varnames": [
                "StatusTodo",
                "StatusInProgress",
                "StatusDone"
            ]
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}